improvements and test page url
