+++ apps/back-office/src/app/dashboard/pages/dashboard/dashboard.component.html
@@ -33,7 +33,7 @@
           (cdkDropListDropped)="onButtonActionDrop($event)"
         >
           <!-- Button actions -->
-          <app-button-action [buttonActions]="buttonActions" [dashboard]="dashboard"></app-button-action>
+          <safe-button-action [buttonActions]="buttonActions" [dashboard]="dashboard"></safe-button-action>
 
           <!-- Fullscreen -->
           <ui-button
