+++ libs/safe/src/lib/components/button-action/button-action-type.ts
@@ -1,0 +1,12 @@
+import { Category, Variant } from '@oort-front/ui';
+
+/**
+ * Button Action Type
+ */
+export type ButtonActionT = {
+  text: string;
+  href: string;
+  variant: Variant;
+  category: Category;
+  openInNewTab: boolean;
+};
