+++ libs/safe/src/lib/components/button-action/public-api.ts
@@ -1,2 +1,3 @@
 export * from './button-action.component';
 export * from './button-action.module';
+export * from './button-action-type';
